Added description in property, fixed decimal type property type

+++ src/constant/businessMetadataTemplate.ts
@@ -144,7 +144,7 @@ export const ATTRIBUTE_TYPES = [
     },
     {
         label: 'Decimal',
-        id: 'decimal',
+        id: 'float',
         icon: 'Number',
     },
     {
